Completed course

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 
@@ -8,13 +8,14 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="This field cannot be blank")
     parser.add_argument('store_id', type=int, required=True, help="This field cannot be blank")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {"message": "Item not found"}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": "An item with name '{}' already exists".format(name)}, 400
@@ -28,7 +29,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "Admin privilege required"}, 500
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -47,5 +53,12 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items": items}, 200
+        return {"items": [item['name'] for item in items],
+                "message": "More data available if you login"
+                }, 200
